Added payment message separation
